Added 2 dockerfiles, DockerfileFull and DockerfileLight
DockerfileLight assumes that yarn build has already been run.
The dockerfile copies the contents of dist into an nginx alpine container.

DockerfileFull assumes that yarn has not been built, and will build and deploy the application.
The dockerfile is based on nginx:stable-alpine.
The dockerfile first installs nodejs and yarn.
After this, it copies the application into a folder. Then, it runs yarn and yarn build.
Lastly, it copies the newly created contents of dist into the nginx folder.